+++ packages/transformer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @orca-fe/painter
 
+## 0.3.15
+
+### Patch Changes
+
+- update dependencies
+
 ## 0.3.14
 
 ### Patch Changes